Clean up some warnings and unused args classes

+++ include/okapi/api/control/util/pidTuner.hpp
@@ -36,8 +36,8 @@ class PIDTuner {
            double ikIMax,
            double ikDMin,
            double ikDMax,
-           std::int32_t inumIterations = 5,
-           std::int32_t inumParticles = 16,
+           std::size_t inumIterations = 5,
+           std::size_t inumParticles = 16,
            double ikSettle = 1,
            double ikITAE = 2);
 
@@ -76,8 +76,8 @@ class PIDTuner {
   const double kIMax;
   const double kDMin;
   const double kDMax;
-  const std::int32_t numIterations;
-  const std::int32_t numParticles;
+  const std::size_t numIterations;
+  const std::size_t numParticles;
   const double kSettle;
   const double kITAE;
 };
